Move create user endpoint to routes directory
