+++ routes/users/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  create: require('./create')
+}
